Move preference code to new preference module
