+++ app/src/main/java/org/jellyfin/androidtv/preference/LibraryPreferences.kt
@@ -3,8 +3,10 @@ package org.jellyfin.androidtv.preference
 import org.jellyfin.androidtv.constant.GridDirection
 import org.jellyfin.androidtv.constant.ImageType
 import org.jellyfin.androidtv.constant.PosterSize
+import org.jellyfin.androidtv.preference.store.DisplayPreferencesStore
 import org.jellyfin.apiclient.model.entities.SortOrder
 import org.jellyfin.apiclient.model.querying.ItemSortBy
+import org.jellyfin.preference.Preference
 import org.jellyfin.sdk.api.client.ApiClient
 
 class LibraryPreferences(
