+++ app/src/main/java/org/jellyfin/androidtv/preference/UserSettingPreferences.kt
@@ -1,6 +1,8 @@
 package org.jellyfin.androidtv.preference
 
 import org.jellyfin.androidtv.constant.HomeSectionType
+import org.jellyfin.androidtv.preference.store.DisplayPreferencesStore
+import org.jellyfin.preference.Preference
 import org.jellyfin.sdk.api.client.ApiClient
 
 class UserSettingPreferences(
