+++ app/src/main/java/org/jellyfin/androidtv/ui/browsing/DisplayPreferencesScreen.kt
@@ -5,12 +5,12 @@ import org.jellyfin.androidtv.constant.GridDirection
 import org.jellyfin.androidtv.constant.ImageType
 import org.jellyfin.androidtv.constant.PosterSize
 import org.jellyfin.androidtv.preference.LibraryPreferences
-import org.jellyfin.androidtv.preference.PreferenceStore
 import org.jellyfin.androidtv.preference.PreferencesRepository
 import org.jellyfin.androidtv.ui.preference.dsl.OptionsFragment
 import org.jellyfin.androidtv.ui.preference.dsl.checkbox
 import org.jellyfin.androidtv.ui.preference.dsl.enum
 import org.jellyfin.androidtv.ui.preference.dsl.optionsScreen
+import org.jellyfin.preference.store.PreferenceStore
 import org.koin.android.ext.android.inject
 
 class DisplayPreferencesScreen : OptionsFragment() {
