+++ app/src/main/java/org/jellyfin/androidtv/ui/preference/dsl/OptionsItemEnum.kt
@@ -3,9 +3,9 @@ package org.jellyfin.androidtv.ui.preference.dsl
 import android.content.Context
 import androidx.annotation.StringRes
 import androidx.preference.PreferenceCategory
-import org.jellyfin.androidtv.preference.Preference
-import org.jellyfin.androidtv.preference.PreferenceStore
 import org.jellyfin.androidtv.ui.preference.custom.RichListPreference
+import org.jellyfin.preference.Preference
+import org.jellyfin.preference.store.PreferenceStore
 import java.util.UUID
 
 class OptionsItemEnum<T : Enum<T>>(
