+++ app/src/main/java/org/jellyfin/androidtv/ui/preference/screen/HomePreferencesScreen.kt
@@ -2,11 +2,11 @@ package org.jellyfin.androidtv.ui.preference.screen
 
 import org.jellyfin.androidtv.R
 import org.jellyfin.androidtv.constant.HomeSectionType
-import org.jellyfin.androidtv.preference.PreferenceStore
 import org.jellyfin.androidtv.preference.UserSettingPreferences
 import org.jellyfin.androidtv.ui.preference.dsl.OptionsFragment
 import org.jellyfin.androidtv.ui.preference.dsl.enum
 import org.jellyfin.androidtv.ui.preference.dsl.optionsScreen
+import org.jellyfin.preference.store.PreferenceStore
 import org.koin.android.ext.android.inject
 
 class HomePreferencesScreen : OptionsFragment() {
