+++ preference/build.gradle.kts
@@ -22,8 +22,21 @@ android {
 		lintConfig = file("$rootDir/android-lint.xml")
 		abortOnError = false
 	}
+
+	testOptions.unitTests.all {
+		it.useJUnitPlatform()
+	}
 }
 
 dependencies {
+	// Kotlin
+	implementation(libs.kotlinx.coroutines)
 
+	// Logging
+	implementation(libs.timber)
+
+	// Testing
+	testImplementation(libs.kotest.runner.junit5)
+	testImplementation(libs.kotest.assertions)
+	testImplementation(libs.mockk)
 }
