+++ preference/src/main/AndroidManifest.xml
@@ -1,0 +1,7 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="org.jellyfin.preference">
+
+    <application>
+
+    </application>
+</manifest>
