+++ preference/src/main/kotlin/store/AsyncPreferenceStore.kt
@@ -1,4 +1,4 @@
-package org.jellyfin.androidtv.preference
+package org.jellyfin.preference.store
 
 import kotlinx.coroutines.runBlocking
